requirements.txt
Flask
typing-extensions
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.44
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langgraph==0.4.3
langgraph-checkpoint==2.0.19
langgraph-checkpoint-sqlite==2.0.5
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.56
langgraph-supervisor==0.0.21
langsmith==0.3.13
langchain-ollama
flask-socketio
ipykernel
google-auth==2.39.0
google-api-python-client==2.173.0